bad fix

+++ pipe.c
@@ -6,7 +6,7 @@
 /*   By: mleclair <mleclair@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/16 13:24:11 by mleclair          #+#    #+#             */
-/*   Updated: 2017/04/29 14:37:12 by mleclair         ###   ########.fr       */
+/*   Updated: 2017/04/29 16:41:37 by mleclair         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,6 +59,8 @@ void		rd_pipe(t_env *env)
 		env->isoperand = 1;
 		frk_pipe(env);
 		env->isoperand = 0;
+		if ((int)getpgid(0) == env->pid)
+			exit(0);
 		kill(0, SIGQUIT);
 	}
 	wait(&status);
